Bugtesting update.
Check it out, see how it works, then work on the form recognition
functions.
